split ALU Class into its own file, add author tags in each file
