+++ ls8/utils.py
@@ -1,3 +1,6 @@
+__author__ = "Stack Overflow"
+
+
 class BColors:
     HEADER = '\033[95m'
     OK_BLUE = '\033[94m'
@@ -8,4 +11,3 @@ class BColors:
     END_ = '\033[0m'
     BOLD = '\033[1m'
     UNDERLINE = '\033[4m'
-
